# target/classes/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Oderstore - OpenAPI 3.0
  description: |-
    This is a sample Oder Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the oder store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Oderstore, then click [here](https://editor.swagger.io/?url=https://oderstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Oderstore OAS 2.0` menu option!_
    
    Some useful links:
    - [The Oder Store repository](https://github.com/swagger-api/swagger-oderstore)
    - [The source API definition for the Oder Store](https://github.com/swagger-api/swagger-oderstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://oderstore3.swagger.io/api/v3
...
paths:
...
  /oder/findByStatus:
    get:
      tags:
        - oder
      summary: Finds Oders by status
      description: Multiple status values can be provided with comma separated strings
      operationId: findOdersByStatus
      parameters:
        - name: status
          in: query
          description: Status values that need to be considered for filter
          required: false
          explode: true
          schema:
            type: string
            default: available
            enum:
              - available
              - pending
              - sold
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Oder'          
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Oder'
        '400':
          description: Invalid status value
      security:
        - orderstore_auth:
            - write:oders
            - read:oders
  /oder/findByTags:
    get:
      tags:
        - oder
      summary: Finds Oders by tags
      description: Multiple tags can be provided with comma separated strings. Use tag1, tag2, tag3 for testing.
      operationId: findOdersByTags
      parameters:
        - name: tags
          in: query
          description: Tags to filter by
          required: false
          explode: true
          schema:
            type: array
            items:
              type: string
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Oder'          
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Oder'
        '400':
          description: Invalid tag value
      security:
        - orderstore_auth:
            - write:oders
            - read:oders
  /oder/{oderId}:
    get:
      tags:
        - oder
      summary: Find oder by ID
      description: Returns a single oder
      operationId: getOderById
      parameters:
        - name: oderId
          in: path
          description: ID of oder to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Oder'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Oder'
        '400':
          description: Invalid ID supplied
        '404':
          description: Oder not found
      security:
        - api_key: []
        - orderstore_auth:
            - write:oders
            - read:oders
    post:
      tags:
        - oder
      summary: Updates a oder in the store with form data
      description: ''
      operationId: updateOderWithForm
      parameters:
        - name: oderId
          in: path
          description: ID of oder that needs to be updated
          required: true
          schema:
            type: integer
            format: int64
        - name: name
          in: query
          description: Name of oder that needs to be updated
          schema:
            type: string
        - name: status
          in: query
          description: Status of oder that needs to be updated
          schema:
            type: string
      responses:
        '400':
          description: Invalid input
      security:
        - orderstore_auth:
            - write:oders
            - read:oders
    delete:
      tags:
        - oder
      summary: Deletes a oder
      description: delete a oder
      operationId: deleteOder
      parameters:
        - name: api_key
          in: header
          description: ''
          required: false
          schema:
            type: string
        - name: oderId
          in: path
          description: Oder id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid oder value
      security:
        - orderstore_auth:
            - write:oders
            - read:oders
  /oder/{oderId}/uploadImage:
    post:
      tags:
        - oder
      summary: uploads an image
      description: ''
      operationId: uploadFile
      parameters:
        - name: oderId
          in: path
          description: ID of oder to update
          required: true
          schema:
            type: integer
            format: int64
        - name: additionalMetadata
          in: query
          description: Additional Metadata
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
        - orderstore_auth:
            - write:oders
            - read:oders
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Dogs
      xml:
        name: category
...
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      xml:
        name: tag
    Oder:
      required:
        - name
        - photoUrls
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: doggie
        category:
          $ref: '#/components/schemas/Category'
        photoUrls:
          type: array
          xml:
            wrapped: true
          items:
            type: string
            xml:
              name: photoUrl
        tags:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/Tag'
        status:
          type: string
          description: oder status in the store
          enum:
            - available
            - pending
            - sold
      xml:
        name: oder
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: '##default'
...
  securitySchemes:
    orderstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://oderstore3.swagger.io/oauth/authorize
          scopes:
            write:oders: modify oders in your account
            read:oders: read your oders
    api_key:
      type: apiKey
      name: api_key
      in: header